+++ sdk/nodejs/core/index.ts
@@ -2,6 +2,7 @@
 // *** Do not edit by hand unless you're certain you know what you are doing! ***
 
 // Export members:
+export * from "./customProvider";
 export * from "./getClientConfig";
 export * from "./getResourceGroup";
 export * from "./getResources";